v0.3.0: Add embed management and container lifecycle features
New features:
- EmbedManager: create, list, get, update, delete embeddable code environments
- 30+ embed templates (React, Python, Flutter, Go, Rust, etc.)
- EmbedSettings with layout presets, themes, analytics, sessions
- Container lifecycle: heartbeat, extend_timeout, hibernate/wake, keep-alive
- LifecycleConfig with presets (development, agent_task, always_on)
- Client patch() method for PATCH requests

Updated:
- Version bumped to 0.3.0
- README with embeds and lifecycle documentation
- CHANGELOG with full v0.3.0 entry
- pyproject.toml: Beta classifier, new keywords, removed setuptools_scm
- .gitignore: exclude test scripts with API keys and internal docs

+++ fleeks_sdk/models.py
@@ -213,7 +213,7 @@ class ContainerStats:
     """
     container_id: str
     cpu_percent: float
-    memory_mb: int
+    memory_mb: float
     memory_percent: float
     network_rx_mb: float
     network_tx_mb: float
